:test_tube: Test: unit tests for signup
