+++ src/repositories/userRepository.ts
@@ -1,7 +1,7 @@
 import prisma from "../database";
 
 async function findByEmail(email: string){
-    return prisma.users.findFirst({
+    return await prisma.users.findFirst({
         where: {
             email
         }
@@ -9,7 +9,7 @@ async function findByEmail(email: string){
 }
 
 async function insert(email: string, password: string){
-    return prisma.users.create({
+    return await prisma.users.create({
         data: {
             email,
             password
